Configurando edicion de datos persona

+++ src/main/java/com/portfolio/rn/Interface/IPersonaService.java
@@ -11,7 +11,8 @@ public interface IPersonaService {
     
     public void savePersona(Persona persona);
     
-    public void deletePersona(Long id);
+    public void deletePersona(int id);
     
-    public Persona findPersona(Long id);
+    public Persona findPersona(int id);
+   
 }
